Fix Compiling Error for Commands.cs
Commands.cs runs the startup and is attatched to the terminal.
Command.cs is the abstract class itself.
StandardCommands.cs holds all the standard commands the user has
when the game starts.
Add commands specific to different objects by making a new
class that derives the abstract Command class, and put that new
class in the object you're hacking.